feat: add AES-128 GCM and test

+++ src/padding.c
@@ -1,7 +1,7 @@
 
 #include "padding.h"
 
-void pkcs7_padding(uint8_t *data, uint32_t *data_size, uint8_t block_size)
+void pkcs7_padding(uint8_t *data, uint64_t *data_size, uint8_t block_size)
 {
         uint8_t padding_size;
         padding_size = block_size - (*data_size % block_size);
@@ -9,7 +9,7 @@ void pkcs7_padding(uint8_t *data, uint32_t *data_size, uint8_t block_size)
         *data_size += padding_size;
 }
 
-void pkcs7_unpadding(uint8_t *data, uint32_t *data_size, uint8_t block_size)
+void pkcs7_unpadding(uint8_t *data, uint64_t *data_size, uint8_t block_size)
 {
         uint8_t final_data = data[*data_size - 1];
         uint32_t padding_idx;
